requirements.txt
fastapi
uvicorn
requests
playwright
pytest
httpx
python-dotenv
litellm
tldextract
